arm64: dts: Add basic support for BIQU CB1
CB1 is Compute Module style board that plugs
into Rpi board style adapter or Manta 3D printer
boards (M4P/M8P).

The board has H616 SoC
1GB of RAM
AXP313A PMIC (Which seems to be very very
close to AXP1530 as it uses AXP1530 driver downstream)
And the actual boards that CB1 plugs in are
just extension to it with ports and thus are not
differenicatedin DT.

Boards have:
4x (3x for Manta boards) USB and 1 USB OTG.
SDcard slot for loading images (Currently not working)
Ethernet port is wired to the internal PHY
and thus only 100Mbps is supported (Currently unsupported upstream).
It also has 2x HDMI 2.0.
And Power and Status LEDs.